# openapi.json
{"openapi":"3.1.0","info":{"title":"AI Endoscopy Clinical Decision Support System API","description":"Agentic AI-based Endoscopy CDSS using Swin Transformer, Grad-CAM, RAG and Clinical Practice Guidelines","version":"1.0.0"},"paths":{"/api/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","description":"Register a new clinician account.","operationId":"signup_api_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Authenticate and return a JWT token.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","description":"Return the authenticated user's profile.","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/patients":{"get":{"tags":["Patients"],"summary":"List Patients","description":"List all patients belonging to the authenticated user.","operationId":"list_patients_api_patients_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PatientResponse"},"type":"array","title":"Response List Patients Api Patients Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Patients"],"summary":"Create Patient","description":"Create a new patient record owned by the authenticated user.","operationId":"create_patient_api_patients_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/patients/{patient_id}":{"get":{"tags":["Patients"],"summary":"Get Patient","description":"Get a specific patient — ownership enforced.","operationId":"get_patient_api_patients__patient_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Patients"],"summary":"Update Patient","description":"Update a patient record — ownership enforced.","operationId":"update_patient_api_patients__patient_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyze":{"post":{"tags":["Analysis"],"summary":"Analyze","description":"POST /api/analyze\nFull clinical analysis pipeline:\n1. Authenticate user\n2. Validate patient ownership\n3. Save uploaded image\n4. Run Swin Transformer prediction + Grad-CAM\n5. Run agentic analysis (Patient Tool + RAG Tool + LLM reasoning)\n6. Save prediction + report to database\n7. Return structured response","operationId":"analyze_api_analyze_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_api_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/patients/{patient_id}/predictions":{"get":{"tags":["Analysis"],"summary":"Get Patient Predictions","description":"Get prediction history for a specific patient — ownership enforced.","operationId":"get_patient_predictions_api_patients__patient_id__predictions_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/{report_id}":{"get":{"tags":["Reports"],"summary":"Get Report","description":"Retrieve full structured report JSON by report_id.","operationId":"get_report_api_reports__report_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"string","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_analyze_api_analyze_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image","description":"Endoscopy image file"},"patient_id":{"type":"integer","title":"Patient Id","description":"Patient ID (integer)"}},"type":"object","required":["image","patient_id"],"title":"Body_analyze_api_analyze_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"PatientCreate":{"properties":{"full_name":{"type":"string","title":"Full Name"},"age":{"type":"integer","title":"Age"},"sex":{"type":"string","title":"Sex"},"medical_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medical History"},"allergies":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allergies"},"medications":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medications"},"previous_endoscopy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous Endoscopy"},"family_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family History"}},"type":"object","required":["full_name","age","sex"],"title":"PatientCreate"},"PatientResponse":{"properties":{"id":{"type":"integer","title":"Id"},"patient_ref":{"type":"string","title":"Patient Ref"},"full_name":{"type":"string","title":"Full Name"},"age":{"type":"integer","title":"Age"},"sex":{"type":"string","title":"Sex"},"medical_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medical History"},"allergies":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allergies"},"medications":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medications"},"previous_endoscopy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous Endoscopy"},"family_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family History"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","patient_ref","full_name","age","sex","created_at","updated_at"],"title":"PatientResponse"},"PatientUpdate":{"properties":{"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"sex":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sex"},"medical_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medical History"},"allergies":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allergies"},"medications":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Medications"},"previous_endoscopy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous Endoscopy"},"family_history":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family History"}},"type":"object","title":"PatientUpdate"},"SignupRequest":{"properties":{"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"confirm_password":{"type":"string","title":"Confirm Password"}},"type":"object","required":["full_name","email","password","confirm_password"],"title":"SignupRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"TokenResponse"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","full_name","email","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}